added regex to collect recon method
There are a handful of examples of reconstruction method strings in tests_dcm2niix4pet.py
and we are going to write methods to handle these few cases that we have examples of.
This is by no means the best way or even the correct way to determine the reconstruction
method from a dicom tag, ultimately it's up to the user/curater to determine what is
correct, this is just a best guess.

ReconstructionIterations and ReconstructionSubsets are denoted with as either by a
single char and followed/proceeded by digits representing the number of iterations
or subsets. get_recon_method is written to extract this iteration_subset string,
determine if iterations come first or if subsets are first listed right to left.

Once that is determined the digits can be counted/recorded using more regex.