Added open_metadata and translate_metadata methods
Added open_metadata and translate_metadata methods to helper methods
while using existing example metadata_excel_example_reader.py to
import metadata from one of our example spreadsheets. The intent
is to create a translation script that will parse a spreadsheet
like file and return the BIDS required fields from it. While
providing the user with the ability to load this script at the
time of conversion, e.g. a user would create a translation script
per study and point to it to in order to extract blood and/or radiological
data from an existing spreadsheet for a given subject(s).

from the command line this would look something like:

`convert image_folder -spreadsheet-path <path to spreadsheet> \
-translation-scrip_path <translation_script.py>`